% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOpenET_multipolygon.R
\name{getOpenET_multipolygon}
\alias{getOpenET_multipolygon}
\title{Timeseries of monthly/daily ET for multiple custom polygons}
\usage{
getOpenET_multipolygon(
  start_date = "2021-01-01",
  end_date = as.character(Sys.Date()),
  model = "ensemble",
  variable = "et",
  units = "in",
  reference_et = "cimis",
  interval = "monthly",
  overpass = "false",
  reducer = "mean",
  asset_id,
  attributes = "",
  api_key
)
}
\arguments{
\item{start_date}{The start date as a string in 'yyyy-mm-dd' format.}

\item{end_date}{The end date as a string in 'yyyy-mm-dd' format. Defaults to today's date.}

\item{model}{The ET model: 'ensemble', 'eemetric', 'ssebop', 'geesebal', 'sims', 'disalexi', 'ptjpl'. Defaults to 'ensemble'.}

\item{variable}{Variable to fetch: 'et', 'et_mad_min', 'et_mad_max', 'eto', 'etr', 'etof', 'ndvi'. Defaults to 'et'.}

\item{units}{Units for ET; 'mm' or 'in' (inches). Defaults to 'in'.}

\item{reference_et}{Reference ET source, either 'cimis' (CA only) or 'gridmet' (all states). Defaults to 'cimis'.}

\item{interval}{Time interval: 'daily' or 'monthly'. Defaults to 'daily'.}

\item{overpass}{Return only data from the satellite overpass days?: 'true' or 'false' as a string. Defaults to 'false'.}

\item{reducer}{Pixel aggregation method for the polygon: 'mean', 'median', 'min', 'max', or 'sum'. Defaults to 'mean'.}

\item{asset_id}{Asset ID for the Google Earth Engine shapefile asset. e.g. 'projects/user/assets/my_shapefile'}

\item{attributes}{Names of shapefile attributes to include in the response csv file, as an R vector of strings.}

\item{api_key}{Your personal OpenET API token as a string.}
}
\value{
Returns a url where the datafile can be downloaded. It may take a few minutes before the file is ready for download. If the url is stored as an R variable, data can then be read in using `read_csv(file = url)`
}
\description{
Makes calls to the OpenET Raster/Timeseries/Multipolygon API endpoint. Polygons must be uploaded to Google Earth Engine and shared with `openet@googlegroups.com`
}
\details{
Note that all parameters are strings. Most parameters have default values and can therefore be omitted,
if the defaults are acceptable.

OpenET API Documentation for this endpoint:

https://openet.gitbook.io/docs/reference/api-reference/raster#timeseries-multipolygon

OpenET API Testbed:

https://openet-api.org/#/Retrieve%20Raster%20Data/_raster_timeseries_multipolygon_post
}
\examples{
getOpenET_multipolygon(start_date = '2020-01-01', end_date = '2021-12-31', model = 'ensemble', units = 'mm', interval = 'monthly', variable = 'eto', reference_et = 'gridmet', asset_id = 'projects/myname/assets/my_shapefile', api_key = mykey)

}
