% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOpenET_polygon.R
\name{getOpenET_polygon}
\alias{getOpenET_polygon}
\title{Timeseries of monthly/daily ET for a custom polygon}
\usage{
getOpenET_polygon(
  geometry,
  start_date = "2021-01-01",
  end_date = as.character(Sys.Date()),
  model = "ensemble",
  variable = "et",
  units = "in",
  reference_et = "cimis",
  interval = "daily",
  reducer = "mean",
  overpass = FALSE,
  api_key = ""
)
}
\arguments{
\item{geometry}{A numeric vector containing longitude-latitude pairs for a single polygon. e.g. `c(-114.2, 33.5, -114.8, 33.7, -114.0, 33.0)`}

\item{start_date}{The start date as a string in 'yyyy-mm-dd' format.}

\item{end_date}{The end date as a string in 'yyyy-mm-dd' format. Defaults to today's date.}

\item{model}{The ET model: 'ensemble', 'eemetric', 'ssebop', 'geesebal', 'sims', 'disalexi', 'ptjpl'. Defaults to 'ensemble'.}

\item{variable}{Variable to fetch: 'et', 'et_mad_min', 'et_mad_max', 'eto', 'etr', 'etof', 'ndvi'. Defaults to 'et'.}

\item{units}{Units for ET; 'mm' or 'in' (inches). Defaults to 'in'.}

\item{reference_et}{Reference ET source, either 'cimis' (CA only) or 'gridmet' (all states). Defaults to 'cimis'.}

\item{interval}{Time interval: 'daily' or 'monthly'. Defaults to 'daily'.}

\item{reducer}{Pixel aggregation method for the polygon: 'mean', 'median', 'min', 'max', or 'sum'. Defaults to 'mean'.}

\item{overpass}{Boolean. If TRUE, fetches data only for Landsat overpass dates. Defaults to FALSE.}

\item{api_key}{Your personal OpenET API token as a string.}
}
\value{
Returns a data frame with one row per time interval and 3 columns:

<date>    Date, in 'yyyy-mm-dd' format

<year>    Numeric year, extracted from <date>

<month>   Numeric month, extracted from <date>

<julian>  Numeric julian day of year, extracted from <date>

<et>      Daily or monthly ET, depending on chosen interval, in either inches or mm, depending on chosen units

<model>   Name of the ET model

<units>   'inches' or 'mm'
}
\description{
Makes calls to the OpenET /raster/timeseries/polygon API endpoint. Use this function when you want
to get data for a single user-defined polygon instead of using OpenET's built-in fields. Data is returned as an R data frame.
}
\details{
If you don't know the lat-long coordinates of your desired polygon, there is an easy way to find it. From the OpenET web app,
click the orange Draw Custom Area button at lower right. Click one of the two polygon tools and draw the polygon on the map.
Click Run Timeseries, then click Copy Shape to Clipboard at the top of the popup. Now you can create a numeric vector containing
the pasted lat/long pairs, e.g. `geom <- c(-114.2, 33.5, -114.8, 33.7, -114.0, 33.0)`. Finally, pass the vector to the
`geometry` parameter of the function.

#' Note that all parameters except `geometry` are strings. Most parameters have default values and can therefore be omitted,
if the defaults are acceptable.

OpenET API Documentation for this endpoint:

https://openet.gitbook.io/docs/reference/api-reference/raster#timeseries-polygon

OpenET API Testbed:

https://openet-api.org/#/Retrieve%20Raster%20Data/_raster_timeseries_polygon_post
}
\examples{
getOpenET_polygon(geometry = c(-114.2, 33.5, -114.8, 33.7, -114.0, 33.0), start_date = '2020-01-01', end_date = '2021-12-31', model = 'ensemble', units = 'mm', interval = 'daily', api_key = mykey)

}
