% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openet-package.R
\docType{package}
\name{openet-package}
\alias{openet}
\alias{openet-package}
\title{openet: Tools for accessing the OpenET API}
\description{
Included are functions for accessing several OpenET endpoints: /timeseries/features/monthly, /raster/timeseries/polygon, and /raster/timeseries/multipolygon
}
\keyword{internal}
