% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOpenET_fields.R
\name{getOpenET_fields}
\alias{getOpenET_fields}
\title{Timeseries of monthly ET for one or more OpenET fields}
\usage{
getOpenET_fields(
  field_ids = "06323746",
  start_date = paste0(year(Sys.Date()), "-01-01"),
  end_date = as.character(Sys.Date() - 14),
  model = "ensemble",
  variable = "et",
  interval = "monthly",
  api_key = ""
)
}
\arguments{
\item{field_ids}{One or more OpenET field ids, as a string or vector of strings. To determine the field id, click on a field within the OpenET web explorer. Id is shown in the popup title.}

\item{start_date}{The start date as a string in 'yyyy-mm-dd' format. Defaults to start of current year.}

\item{end_date}{The end date as a string in 'yyyy-mm-dd' format. Defaults to two weeks ago.}

\item{model}{A vector of one of more ET models to use: 'ensemble', 'eemetric', 'ssebop', 'geesebal', 'sims', 'disalexi', 'ptjpl'. Defaults to 'ensemble'.}

\item{variable}{A vector of one or more variables to fetch: 'et', 'et_mad_min', 'et_mad_max', 'eto', 'etr', 'etof', 'ndvi', 'pr'. Defaults to 'et'.}

\item{interval}{Time interval: 'daily' or 'monthly'. Defaults to 'monthly'.}

\item{api_key}{Your personal OpenET API token as a string.}

\item{units}{Units for ET; 'mm' or 'in' (inches). Defaults to 'in'.}
}
\value{
Returns a data frame with (months) x (fields) rows and 5 columns:

<start_date>  Start date of the month, in 'yyyy-mm-dd' format

<end_date>    End date of the month

<month>       Numeric month, extracted from <start_date>

<year>        Numeric year, extracted from <start_date>

<field>       OpenET field id

<et>          Total monthly ET (sum of all daily values) in either inches or mm, depending on chosen units

<units>       'inches' or 'mm'
}
\description{
Makes calls to the OpenET /timeseries/features/monthly API endpoint. Use this function when you
want to use OpenET's built-in field boundaries, and you know the field id's of the specific field(s). The field id's
can be found by clicking on fields in the OpenET web app and looking at the top of the resulting popup. Data is
returned as an R data frame.
}
\details{
Note that all parameters are strings, esp. field id's which may have leading zeroes. Most parameters have default values
and can therefore be omitted if the defaults are acceptable.

OpenET API Documentation:

https://open-et.github.io/docs/build/html/geo_timeseries.html#timeseries-features-monthly

OpenET API Testbed:

https://openet.dri.edu/docs#/timeseries/feature_timeseries_annual_timeseries_features_monthly_post
}
\examples{
getOpenET_fields(field_ids = c('06323746', '06435895'), '2020-01-01', '2021-12-31', 'ensemble', 'et', 'metric', mykey)
getOpenET_fields(field_ids = '065746833', start_date = '2021-01-01', end_date = '2022-12-31', api_key = mykey)

}
