% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOpenET_fields.R
\name{getOpenET_fields}
\alias{getOpenET_fields}
\title{Timeseries of monthly ET for one or more OpenET fields}
\usage{
getOpenET_fields(
  state = "CA",
  field_ids = "06323746",
  start_date = "2021-01-01",
  end_date = as.character(Sys.Date() - 14),
  model = "ensemble_mean",
  variable = "et",
  units = "english",
  api_key = ""
)
}
\arguments{
\item{state}{Two-letter state abbreviation, e.g. 'CA', 'AZ'. Defaults to 'CA'.}

\item{field_ids}{One or more OpenET field ids, as a string or vector of strings. To determine the field id, click on a field within the OpenET web explorer. Id is shown in the popup title.}

\item{start_date}{The start date as a string in 'yyyy-mm-dd' format.}

\item{end_date}{The end date as a string in 'yyyy-mm-dd' format. Defaults to two weeks ago.}

\item{model}{ET model to use: 'ensemble_mean', 'eemetric', 'ssebop', 'geesebal', 'sims', 'disalexi', 'ptjpl'. Defaults to 'ensemble_mean'.}

\item{variable}{Variable to fetch: 'et', 'ndvi', 'et_reference', 'et_fraction', 'count'. Defaults to 'et'.}

\item{units}{Units for ET; 'metric' will give mm and 'english' will give inches. Defaults to 'english'.}

\item{api_key}{Your personal OpenET API key as a string. Can be obtained from https://auth.etdata.org}
}
\value{
Returns a data frame with (months) x (fields) rows and 5 columns:

<start_date>  Start date of the month

<end_date>    End date of the month

<field>       OpenET field id

<et>          Total monthly ET (sum of all daily values) in either inches or mm, depending on chosen units

<units>       'inches' or 'mm'
}
\description{
`getOpenET_fields` makes calls to the OpenET /timeseries/features/monthly API endpoint. Use this function when you
want to use OpenET's built-in field boundaries, and you know the field id's of the specific field(s). The field id's
can be found by clicking on fields in the OpenET web app and looking at the top of the resulting popup. Data is
returned as an R data frame.
}
\details{
Note that all parameters are strings, esp. field id's which may have leading zeroes. Most parameters have default values
and can therefore be omitted if the defaults are acceptable.

OpenET API Documentation:

https://open-et.github.io/docs/build/html/geo_timeseries.html#timeseries-features-monthly

OpenET API Testbed:

https://openet.dri.edu/docs#/timeseries/feature_timeseries_annual_timeseries_features_monthly_post
}
\examples{
getOpenET_fields('AZ', c('06323746', '06435895'), '2020-01-01', '2021-12-31', 'ensemble_mean', 'et', 'metric', 'mykey')
getOpenET_fields(field_ids = '065746833', start_date = '2021-01-01', end_date = '2022-12-31', api_key = 'mykey')
}
