% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOpenET_multipolygon.R
\name{getOpenET_multipolygon}
\alias{getOpenET_multipolygon}
\title{Timeseries of monthly/daily ET for a multiple custom polygons}
\usage{
getOpenET_multipolygon(
  start_date = "2020-01-01",
  end_date = as.character(Sys.Date()),
  model = "eemetric",
  variable = "et",
  ref_et_source = "cimis",
  units = "english",
  interval = "monthly",
  shapefile_asset_id = "projects/lsteely/assets/MWD_Parcels",
  include_columns = "PVID_PARNU,WT_ACRES",
  output_file_format = "csv",
  filename_suffix = "ls",
  api_key = ""
)
}
\arguments{
\item{start_date}{The start date in yyyy-mm-dd format.}

\item{end_date}{The end date in yyyy-mm-dd format. Defaults to today's date.}

\item{model}{'ensemble_mean', 'eemetric', 'ssebop', 'geesebal', 'sims', 'disalexi', 'ptjpl'. Defaults to 'eemetric'.}

\item{variable}{'et', 'ndvi', 'et_reference', 'et_fraction', 'count'. Defaults to 'et'.}

\item{ref_et_source}{Reference ET source, either 'cimis' (CA only) or 'gridmet' (all states). Defaults to 'cimis'.}

\item{units}{'metric', 'english'. Defaults to 'english'.}

\item{interval}{'daily', 'monthly'. Defaults to 'monthly'.}

\item{shapefile_asset_id}{Path to an Earth Engine shapefile asset containing the polygons.}

\item{include_columns}{Additional (non-OpenET) columns from the shapefile to include in the resulting data frame.}

\item{output_file_format}{'geojson', 'csv'. Defaults to 'csv'.}

\item{filename_suffix}{String to append to filename, for identification.}

\item{api_key}{OpenET API key (long string), obtain from https://auth.etdata.org}
}
\value{
A string containing the url from which the data can be downloaded as a .csv file.
}
\description{
`getOpenET_multipolygon` makes calls to the OpenET Raster/Timeseries/Multipolygon API endpoint.
}
\details{
OpenET API documentation:

https://open-et.github.io/docs/build/html/ras_timeseries.html#raster-timeseries-multipolygon

OpenET API Testbed:

https://openet.dri.edu/docs#/raster/raster_timeseries_multipolygon_raster_timeseries_multipolygon_get
}
\examples{
getOpenET_multipolygon(start_date = '2021-01-01', end_date = '2021-12-31', model = 'ensemble_mean', units = 'mm', interval = 'daily', shapefile_asset_id = 'projects/penman/my_shapefile', output_file_format = 'csv', api_key = 'mykey')
}
