% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOpenET_polygon.R
\name{getOpenET_polygon}
\alias{getOpenET_polygon}
\title{Timeseries of monthly/daily ET for a custom polygon}
\usage{
getOpenET_polygon(
  geometry,
  start_date = "2021-01-01",
  end_date = as.character(Sys.Date()),
  model = "ensemble",
  variable = "et",
  units = "english",
  ref_et_source = "cimis",
  provisional = "false",
  interval = "daily",
  moving_average = "0",
  best_effort = "true",
  pixel_aggregation = "mean",
  api_key = ""
)
}
\arguments{
\item{geometry}{A numeric vector containing lat-long pairs. e.g. `c(-114.2, 33.5, -114.8, 33.7, -114.0, 33.0)`}

\item{start_date}{The start date as a string in 'yyyy-mm-dd' format.}

\item{end_date}{The end date as a string in 'yyyy-mm-dd' format. Defaults to today's date.}

\item{model}{The ET model: 'ensemble', 'eemetric', 'ssebop', 'geesebal', 'sims', 'disalexi', 'ptjpl'. Defaults to 'ensemble_mean'.}

\item{variable}{Variable to fetch: 'et', 'ndvi', 'et_reference', 'et_fraction', 'count'. Defaults to 'et'.}

\item{units}{Units for ET; 'metric' will give mm and 'english' will give inches. Defaults to 'english'.}

\item{ref_et_source}{Reference ET source, either 'cimis' (CA only) or 'gridmet' (all states). Defaults to 'cimis'.}

\item{provisional}{Include most recent months of data, even if provisional? Defaults to 'true'.}

\item{interval}{Time interval: 'daily' or 'monthly'. Defaults to 'daily'.}

\item{moving_average}{Average the data by x moving time intervals. '0', '1', '2', '3', etc. Defaults to '0'.}

\item{best_effort}{Use non-pixel-based interpolation for faster daily data? Defaults to 'true.'}

\item{pixel_aggregation}{Statistical spatial summary: 'mean', 'median', 'min', 'max', 'sum', 'count'. Defaults to 'mean'.}

\item{api_key}{Your personal OpenET API key as a string. OCan be obtained from https://auth.etdata.org}
}
\value{
Returns a data frame with one row per time interval and 3 columns:

<date>    Date, in 'yyyy-mm-dd' format

<year>    Numeric year, extracted from <date>

<month>   Numeric month, extracted from <date>

<julian>  Numeric julian day of year, extracted from <date>

<et>      Daily or monthly ET, depending on chosen interval, in either inches or mm, depending on chosen units

<units>   'inches' or 'mm'
}
\description{
`getOpenET_polygon` makes calls to the OpenET /raster/timeseries/polygon API endpoint. Use this function when you want
to get data for a single user-defined polygon instead of using OpenET's built-in fields. Data is returned as an R data frame.
}
\details{
If you don't know the lat-long coordinates of your desired polygon, there is an easy way to find it. From the OpenET web app,
click the orange Draw Custom Area button at lower right. Click one of the two polygon tools and draw the polygon on the map.
Click Run Timeseries, then click Copy Shape to Clipboard at the top of the popup. Now you can create a numeric vector containing
the pasted lat/long pairs, e.g. `geom <- c(-114.2, 33.5, -114.8, 33.7, -114.0, 33.0)`. Finally, pass the vector to the
`geometry` parameter of the function.

#' Note that all parameters except `gometry` are strings. Most parameters have default values and can therefore be omitted,
if the defaults are acceptable.

OpenET API Documentation:

https://open-et.github.io/docs/build/html/ras_timeseries.html#raster-timeseries-polygon

OpenET API Testbed:

https://openet.dri.edu/docs#/raster/raster_timeseries_polygon_raster_timeseries_polygon_post
}
\examples{
getOpenET_polygon(geometry = c(-114.2, 33.5, -114.8, 33.7, -114.0, 33.0), start_date = '2020-01-01', end_date = '2021-12-31', model = 'ensemble_mean', units = 'metric', interval = 'daily', api_key = 'mykey')

}
